etl/pyproject.toml
[tool.poetry]
name = "01_sprint_03 ETL"
version = "0.1.0"
description = ""
authors = ["captainKoala"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
psycopg2-binary = "^2.9.3"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
